+++ connect/models/schemas.py
@@ -10,6 +10,11 @@ import six
 class BaseSchema(Schema):
     id = fields.Str()
 
+    # Set allow_none to True in all fields
+    def on_bind_field(self, field_name, field_obj):
+        super(BaseSchema, self).on_bind_field(field_name, field_obj)
+        field_obj.allow_none = True
+
     @post_load
     def make_object(self, data):
         from connect.models import BaseModel
@@ -17,9 +22,9 @@ class BaseSchema(Schema):
 
 
 class ActivationSchema(BaseSchema):
-    link = fields.Str(allow_none=True)
+    link = fields.Str()
     message = fields.Str()
-    date = fields.DateTime(allow_none=True)
+    date = fields.DateTime()
 
     @post_load
     def make_object(self, data):
@@ -28,7 +33,7 @@ class ActivationSchema(BaseSchema):
 
 
 class AgreementStatsSchema(BaseSchema):
-    contracts = fields.Int(allow_none=True)
+    contracts = fields.Int()
     versions = fields.Int()
 
     @post_load
@@ -47,10 +52,10 @@ class CompanySchema(BaseSchema):
 
 
 class PhoneNumberSchema(BaseSchema):
-    country_code = fields.Str(allow_none=True)
-    area_code = fields.Str(allow_none=True)
-    phone_number = fields.Str(allow_none=True)
-    extension = fields.Str(allow_none=True)
+    country_code = fields.Str()
+    area_code = fields.Str()
+    phone_number = fields.Str()
+    extension = fields.Str()
 
     @post_load
     def make_object(self, data):
@@ -60,8 +65,8 @@ class PhoneNumberSchema(BaseSchema):
 
 class ContactSchema(BaseSchema):
     email = fields.Str()
-    first_name = fields.Str(allow_none=True)
-    last_name = fields.Str(allow_none=True)
+    first_name = fields.Str()
+    last_name = fields.Str()
     phone_number = fields.Nested(PhoneNumberSchema)
 
     @post_load
@@ -72,7 +77,7 @@ class ContactSchema(BaseSchema):
 
 class ContactInfoSchema(BaseSchema):
     address_line1 = fields.Str()
-    address_line2 = fields.Str(allow_none=True)
+    address_line2 = fields.Str()
     city = fields.Str()
     contact = fields.Nested(ContactSchema)
     country = fields.Str()
@@ -85,7 +90,7 @@ class ContactInfoSchema(BaseSchema):
         return ContactInfo(**data)
 
 
-class ValueChoiceSchema(Schema):
+class ValueChoiceSchema(BaseSchema):
     value = fields.Str()
     label = fields.Str()
 
@@ -130,6 +135,7 @@ class DownloadLinkSchema(BaseSchema):
 
 class UserSchema(BaseSchema):
     name = fields.Str()
+    email = fields.Str()
 
     @post_load
     def make_object(self, data):
@@ -138,8 +144,8 @@ class UserSchema(BaseSchema):
 
 
 class EventSchema(BaseSchema):
-    at = fields.DateTime(allow_none=True)
-    by = fields.Nested(UserSchema, allow_none=True)
+    at = fields.DateTime()
+    by = fields.Nested(UserSchema)
 
     @post_load
     def make_object(self, data):
@@ -188,7 +194,7 @@ class HubStatsSchema(BaseSchema):
 class HubSchema(BaseSchema):
     name = fields.Str()
     company = fields.Nested(CompanySchema)
-    description = fields.Str(allow_none=True)
+    description = fields.Str()
     instance = fields.Nested(HubInstanceSchema)
     events = fields.Nested(EventsSchema)
     stats = fields.Nested(HubStatsSchema)
@@ -199,9 +205,19 @@ class HubSchema(BaseSchema):
         return Hub(**data)
 
 
-class ExtIdHubSchema(Schema):
+class ExternalIdField(fields.Field):
+    def _deserialize(self, value, attr, obj, **kwargs):
+        if isinstance(value, six.string_types):
+            return value
+        elif isinstance(value, int):
+            return str(value)
+        else:
+            raise ValueError({attr: [u'Not a valid int or string.']})
+
+
+class ExtIdHubSchema(BaseSchema):
     hub = fields.Nested(HubSchema, only=('id', 'name'))
-    external_id = fields.Str()
+    external_id = ExternalIdField()
 
     @post_load
     def make_object(self, data):
@@ -256,22 +272,34 @@ class MarketplaceSchema(BaseSchema):
         return Marketplace(**data)
 
 
+class CountrySchema(BaseSchema):
+    name = fields.Str()
+    icon = fields.Str()
+    zone = fields.Str()
+
+    @post_load
+    def make_object(self, data):
+        from connect.models import Country
+        return Country(**data)
+
+
 class ParamSchema(BaseSchema):
     name = fields.Str()
     description = fields.Str()
     type = fields.Str()
-    value = fields.Str(allow_none=True)
-    value_error = fields.Str(allow_none=True)
-    value_choice = fields.Str(many=True, allow_none=True)
+    value = fields.Str()
+    value_error = fields.Str()
+    value_choice = fields.Str(many=True)
 
     # Undocumented fields (they appear in PHP SDK)
-    title = fields.Str(allow_none=True)
-    scope = fields.Str(allow_none=True)
-    constraints = fields.Nested(ConstraintsSchema, allow_none=True)
-    value_choices = fields.Nested(ValueChoiceSchema, many=True, allow_none=True)
-    phase = fields.Str(allow_none=True)
-    events = fields.Nested(EventsSchema, allow_none=True)
-    marketplace = fields.Nested(MarketplaceSchema, allow_none=True)
+    title = fields.Str()
+    scope = fields.Str()
+    constraints = fields.Nested(ConstraintsSchema)
+    value_choices = fields.Nested(ValueChoiceSchema, many=True)
+    phase = fields.Str()
+    events = fields.Nested(EventsSchema)
+    marketplace = fields.Nested(MarketplaceSchema)
+    countries = fields.Nested(CountrySchema, many=True)
 
     @post_load
     def make_object(self, data):
@@ -282,15 +310,15 @@ class ParamSchema(BaseSchema):
 class ItemSchema(BaseSchema):
     mpn = fields.Str()
     quantity = QuantityField()
-    old_quantity = QuantityField(allow_none=True)
-    renewal = fields.Nested(RenewalSchema, allow_none=True)
-    params = fields.Nested(ParamSchema, many=True, allow_none=True)
-    display_name = fields.Str(allow_none=True)
-    global_id = fields.Str(allow_none=True)
-    item_type = fields.Str(allow_none=True)
-    period = fields.Str(allow_none=True)
-    type = fields.Str(allow_none=True)
-    name = fields.Str(allow_none=True)
+    old_quantity = QuantityField()
+    renewal = fields.Nested(RenewalSchema)
+    params = fields.Nested(ParamSchema, many=True)
+    display_name = fields.Str()
+    global_id = fields.Str()
+    item_type = fields.Str()
+    period = fields.Str()
+    type = fields.Str()
+    name = fields.Str()
 
     @post_load
     def make_object(self, data):
@@ -305,16 +333,17 @@ class AgreementSchema(BaseSchema):
     created = fields.DateTime()
     updated = fields.DateTime()
     owner = fields.Nested(CompanySchema)
-    stats = fields.Nested(AgreementStatsSchema, allow_none=True)
-    author = fields.Nested(UserSchema, allow_none=True)
+    stats = fields.Nested(AgreementStatsSchema)
+    author = fields.Nested(UserSchema)
     version = fields.Int()
     active = fields.Bool()
     link = fields.Str()
     version_created = fields.DateTime()
     version_contracts = fields.Int()
     agreements = fields.Nested('AgreementSchema', many=True)
-    parent = fields.Nested('AgreementSchema', only=('id', 'name'), allow_none=True)
-    marketplace = fields.Nested(MarketplaceSchema, only=('id', 'name'), allow_none=True)
+    parent = fields.Nested('AgreementSchema', only=('id', 'name'))
+    marketplace = fields.Nested(MarketplaceSchema, only=('id', 'name'))
+    name = fields.Str()
 
     @post_load
     def make_object(self, data):
@@ -328,15 +357,15 @@ class ContractSchema(BaseSchema):
     type = fields.Str()
     status = fields.Str()
     agreement = fields.Nested(AgreementSchema, only=('id', 'name'))
-    marketplace = fields.Nested(MarketplaceSchema, only=('id', 'name'), allow_none=True)
-    owner = fields.Nested(CompanySchema, only=('id', 'name'), allow_none=True)
+    marketplace = fields.Nested(MarketplaceSchema, only=('id', 'name'))
+    owner = fields.Nested(CompanySchema, only=('id', 'name'))
     creator = fields.Nested(UserSchema, only=('id', 'name'))
     created = fields.DateTime()
     updated = fields.DateTime()
-    enrolled = fields.DateTime(allow_none=True)
+    enrolled = fields.DateTime()
     version_created = fields.DateTime()
     activation = fields.Nested(ActivationSchema)
-    signee = fields.Nested(UserSchema, only=('id', 'name'), allow_none=True)
+    signee = fields.Nested(UserSchema, only=('id', 'name'))
 
     @post_load
     def make_object(self, data):
@@ -377,9 +406,9 @@ class ProductFamilySchema(BaseSchema):
 
 class ProductCategorySchema(BaseSchema):
     name = fields.Str()
-    parent = fields.Nested('ProductCategorySchema', allow_none=True)
-    children = fields.Nested('ProductCategorySchema', many=True, allow_none=True)
-    family = fields.Nested(ProductFamilySchema, allow_none=True)
+    parent = fields.Nested('ProductCategorySchema')
+    children = fields.Nested('ProductCategorySchema', many=True)
+    family = fields.Nested(ProductFamilySchema)
 
     @post_load
     def make_object(self, data):
@@ -408,19 +437,33 @@ class ProductStatsSchema(BaseSchema):
         return ProductStats(**data)
 
 
+class ProductConfigurationParameterSchema(BaseSchema):
+    value = fields.Str()
+    parameter = fields.Nested(ParamSchema)
+    marketplace = fields.Nested(MarketplaceSchema)
+    item = fields.Nested(ItemSchema)
+    events = fields.Nested(EventsSchema)
+    constraints = fields.Nested(ConstraintsSchema)
+
+    @post_load
+    def make_object(self, data):
+        from connect.models import ProductConfigurationParameter
+        return ProductConfigurationParameter(**data)
+
+
 class ProductSchema(BaseSchema):
     name = fields.Str()
     icon = fields.Str()
     short_description = fields.Str()
     detailed_description = fields.Str()
     version = fields.Int()
-    published_at = fields.DateTime(allow_none=True)
+    published_at = fields.DateTime()
     configurations = fields.Nested(ProductConfigurationSchema)
     customer_ui_settings = fields.Nested(CustomerUiSettingsSchema)
-    category = fields.Nested(ProductCategorySchema, allow_none=True)
-    owner = fields.Nested(CompanySchema, allow_none=True)
-    latest = fields.Bool(allow_none=True)
-    stats = fields.Nested(ProductStatsSchema, allow_none=True)
+    category = fields.Nested(ProductCategorySchema)
+    owner = fields.Nested(CompanySchema)
+    latest = fields.Bool()
+    stats = fields.Nested(ProductStatsSchema)
 
     @post_load
     def make_object(self, data):
@@ -428,9 +471,9 @@ class ProductSchema(BaseSchema):
         return Product(**data)
 
 
-class ServerErrorResponseSchema(Schema):
+class ServerErrorResponseSchema(BaseSchema):
     error_code = fields.Str()
-    params = fields.Dict(allow_none=True)
+    params = fields.Dict()
     errors = fields.List(fields.Str())
 
     @post_load
@@ -442,6 +485,7 @@ class ServerErrorResponseSchema(Schema):
 class TemplateSchema(BaseSchema):
     name = fields.Str()
     representation = fields.Str()
+    body = fields.Str()
 
     @post_load
     def make_object(self, data):
@@ -461,7 +505,7 @@ class ConfigurationSchema(BaseSchema):
 class TierAccountSchema(BaseSchema):
     name = fields.Str()
     contact_info = fields.Nested(ContactInfoSchema)
-    external_id = fields.Str()
+    external_id = ExternalIdField()
     external_uid = fields.Str()
 
     @post_load
@@ -470,7 +514,7 @@ class TierAccountSchema(BaseSchema):
         return TierAccount(**data)
 
 
-class TierAccountsSchema(Schema):
+class TierAccountsSchema(BaseSchema):
     customer = fields.Nested(TierAccountSchema)
     tier1 = fields.Nested(TierAccountSchema)
     tier2 = fields.Nested(TierAccountSchema)
@@ -496,15 +540,20 @@ class ConnectionSchema(BaseSchema):
 
 class AssetSchema(BaseSchema):
     status = fields.Str()
-    external_id = fields.Str()
-    external_uid = fields.Str(allow_none=True)
+    external_id = ExternalIdField()
+    events = fields.Nested(EventsSchema)
+    external_uid = fields.Str()
+    external_name = fields.Str()
     product = fields.Nested(ProductSchema, only=('id', 'name'))
     connection = fields.Nested(
-        ConnectionSchema, only=('id', 'type', 'provider', 'vendor'),
+        ConnectionSchema, only=('id', 'type', 'provider', 'vendor')
     )
-    items = fields.Nested(ItemSchema, many=True)
+    contract = fields.Nested(ContractSchema)
+    marketplace = fields.Nested(MarketplaceSchema)
     params = fields.Nested(ParamSchema, many=True)
     tiers = fields.Nested(TierAccountsSchema)
+    items = fields.Nested(ItemSchema, many=True)
+    configuration = fields.Nested(ConfigurationSchema)
 
     @post_load
     def make_object(self, data):
@@ -512,18 +561,28 @@ class AssetSchema(BaseSchema):
         return Asset(**data)
 
 
+class AssigneeField(fields.Field):
+    def _deserialize(self, value, attr, obj, **kwargs):
+        from connect.models.user import User
+        if isinstance(value, six.string_types):
+            return value
+        else:
+            return User.deserialize_json(value)
+
+
 class FulfillmentSchema(BaseSchema):
-    activation_key = fields.Str()
-    asset = fields.Nested(AssetSchema)
-    status = fields.Str()
     type = fields.Str()
-    updated = fields.DateTime()
     created = fields.DateTime()
+    updated = fields.DateTime()
+    status = fields.Str()
+    params_form_url = fields.Str()
+    activation_key = fields.Str()
     reason = fields.Str()
     note = fields.Str()
-    params_form_url = fields.Str()
+    asset = fields.Nested(AssetSchema)
     contract = fields.Nested(ContractSchema, only=('id', 'name'))
     marketplace = fields.Nested(MarketplaceSchema, only=('id', 'name'))
+    assignee = AssigneeField()
 
     @post_load
     def make_object(self, data):
@@ -538,12 +597,13 @@ class TierConfigSchema(BaseSchema):
     tier_level = fields.Int()
     params = fields.Nested(ParamSchema, many=True)
     connection = fields.Nested(ConnectionSchema)
-    open_request = fields.Nested(BaseSchema, allow_none=True)
+    open_request = fields.Nested(BaseSchema)
     template = fields.Nested(TemplateSchema)
     contract = fields.Nested(ContractSchema)
     marketplace = fields.Nested(MarketplaceSchema)
-    configuration = fields.Nested(ConfigurationSchema, allow_none=True)
-    events = fields.Nested(EventsSchema, allow_none=True)
+    configuration = fields.Nested(ConfigurationSchema)
+    events = fields.Nested(EventsSchema)
+    status = fields.Str()
 
     @post_load
     def make_object(self, data):
@@ -561,12 +621,15 @@ class TierConfigRequestSchema(BaseSchema):
     tier_level = fields.Int()
     params = fields.Nested(ParamSchema, many=True)
     environment = fields.Str()
-    assignee = fields.Nested(UserSchema, allow_none=True)
-    template = fields.Nested(TemplateSchema, allow_none=True)
-    reason = fields.Str(allow_none=True)
-    activation = fields.Nested(ActivationSchema, allow_none=True)
-    notes = fields.Str(allow_none=True)
-    events = fields.Nested(EventsSchema, allow_none=True)
+    assignee = fields.Nested(UserSchema)
+    template = fields.Nested(TemplateSchema)
+    reason = fields.Str()
+    activation = fields.Nested(ActivationSchema)
+    notes = fields.Str()
+    events = fields.Nested(EventsSchema)
+    tiers = fields.Nested(TierAccountsSchema)
+    marketplace = fields.Nested(MarketplaceSchema)
+    contract = fields.Nested(ContractSchema)
 
     @post_load
     def make_object(self, data):