+++ connect/models/tier_accounts.py
@@ -3,6 +3,8 @@
 # This file is part of the Ingram Micro Cloud Blue Connect SDK.
 # Copyright (c) 2019 Ingram Micro. All Rights Reserved.
 
+from typing import Optional
+
 from .base import BaseModel
 from .tier_account import TierAccount
 from .schemas import TierAccountsSchema
@@ -19,5 +21,5 @@ class TierAccounts(BaseModel):
     tier1 = None  # type: TierAccount
     """ (:py:class:`.TierAccount`) Level 1 TierAccount Object. """
 
-    tier2 = None  # type: TierAccount
-    """ (:py:class:`.TierAccount`) Level 2 TierAccount Object. """
+    tier2 = None  # type: Optional[TierAccount]
+    """ (:py:class:`.TierAccount` | None) Level 2 TierAccount Object. """