+++ connect/resources/usage_file_automation.py
@@ -3,11 +3,13 @@
 # This file is part of the Ingram Micro Cloud Blue Connect SDK.
 # Copyright (c) 2019 Ingram Micro. All Rights Reserved.
 
+import json
+import logging
 from abc import ABCMeta
 
 from connect.exceptions import SkipRequest, UsageFileAction
-from connect.logger import logger
-from connect.models import BaseModel, UsageFile
+from connect.models.base import BaseModel
+from connect.models.usage_file import UsageFile
 from .automation_engine import AutomationEngine
 
 
@@ -20,6 +22,7 @@ class UsageFileAutomation(AutomationEngine):
     __metaclass__ = ABCMeta
     resource = 'usage/files'
     model_class = UsageFile
+    logger = logging.getLogger('UsageFile.logger')
 
     def filters(self, status='ready', **kwargs):
         """
@@ -30,11 +33,14 @@ class UsageFileAutomation(AutomationEngine):
         """
         filters = super(UsageFileAutomation, self).filters(status, **kwargs)
         if self.config.products:
-            filters['product__id'] = ','.join(self.config.products)
+            filters['product_id'] = ','.join(self.config.products)
         return filters
 
     def dispatch(self, request):
         # type: (UsageFile) -> str
+
+        self._set_custom_logger(request.id, request.name)
+
         try:
             # Validate product
             if self.config.products \
@@ -42,7 +48,7 @@ class UsageFileAutomation(AutomationEngine):
                 return 'Invalid product'
 
             # Process request
-            logger.info(
+            self.logger.info(
                 'Start usage file request process / ID request - {}'.format(request.id))
             result = self.process_request(request)
 
@@ -50,22 +56,22 @@ class UsageFileAutomation(AutomationEngine):
             processing_result = 'UsageFileAutomation.process_request returned {} while ' \
                                 'is expected to raise UsageFileAction or SkipRequest exception' \
                 .format(str(result))
-            logger.warning(processing_result)
+            self.logger.warning(processing_result)
             raise UserWarning(processing_result)
 
         # Catch action
         except UsageFileAction as usage:
             self._api.post(
                 path='{}/{}'.format(request.id, usage.code),
-                json=usage.obj.json
-                if isinstance(usage.obj, BaseModel)
-                else getattr(usage.obj, '__dict__', str(usage.obj)))
+                data=json.dumps(usage.obj.json
+                                if isinstance(usage.obj, BaseModel)
+                                else usage.obj))
             processing_result = usage.code
 
         # Catch skip
         except SkipRequest:
             processing_result = 'skip'
 
-        logger.info('Finished processing of usage file with ID {} with result {}'
-                    .format(request.id, processing_result))
+        self.logger.info('Finished processing of usage file with ID {} with result {}'
+                         .format(request.id, processing_result))
         return processing_result